+++ odbc-api/src/handles/column_description.rs
@@ -2,7 +2,7 @@ use super::data_type::DataType;
 use odbc_sys::WChar;
 use std::char::{decode_utf16, DecodeUtf16Error};
 
-/// Indication of wether a column is nullable or not.
+/// Indication of whether a column is nullable or not.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum Nullable {
     Unknown,
@@ -19,11 +19,11 @@ impl Default for Nullable {
 /// Describes the type and attributes of a column.
 #[derive(Clone, Debug, Eq, PartialEq, Default)]
 pub struct ColumnDescription {
-    /// Colmun name. May be empty if unavailable.
+    /// Column name. May be empty if unavailable.
     pub name: Vec<WChar>,
     /// Type of the column
     pub data_type: DataType,
-    /// Indicates wether the column is nullable or not.
+    /// Indicates whether the column is nullable or not.
     pub nullable: Nullable,
 }
 
@@ -34,7 +34,7 @@ impl ColumnDescription {
         decode_utf16(self.name.iter().copied()).collect()
     }
 
-    /// `true` if the column is `Nullable` or it is not know wether the column is nullable. `false`
+    /// `true` if the column is `Nullable` or it is not know whether the column is nullable. `false`
     /// if and only if the column is `NoNulls`.
     pub fn could_be_nullable(&self) -> bool {
         match self.nullable {