+++ odbc-api/src/handles/logging.rs
@@ -1,7 +1,7 @@
 use super::{as_handle::AsHandle, diagnostics::Record};
 use log::warn;
 
-/// This function inspectas all the diagnostics of an ODBC handle and logs their text messages. It
+/// This function inspects all the diagnostics of an ODBC handle and logs their text messages. It
 /// is going to print placeholder characters, if it cannot convert the message to UTF-8.
 pub fn log_diagnostics(handle: &dyn AsHandle) {
     let mut rec_number = 1;